// client/src/api/model/criticalInjurySeverity.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type CriticalInjurySeverity =
  (typeof CriticalInjurySeverity)[keyof typeof CriticalInjurySeverity];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CriticalInjurySeverity = {
  Easy: "Easy",
  Average: "Average",
  Hard: "Hard",
  Daunting: "Daunting",
  Formidable: "Formidable",
} as const;